requirements.txt
pandas>=2.2,<3.0
numpy>=1.26,<3.0
duckdb>=1.2,<2.0
pyarrow>=16.0,<20.0
openpyxl>=3.1,<4.0
python-dotenv>=1.0,<2.0

# Notebook / visualization
jupyter>=1.0,<2.0
ipykernel>=6.0,<7.0
matplotlib>=3.8,<4.0
seaborn>=0.13,<1.0

# Forecasting / ML
scikit-learn>=1.5,<2.0
statsmodels>=0.14,<1.0
prophet>=1.1,<2.0
xgboost>=2.1,<3.0
pmdarima>=2.0,<3.0

# Orchestration / tracking / app
fredapi>=0.5,<1.0
mlflow>=2.20,<3.0
streamlit>=1.44,<2.0
plotly>=5.24,<6.0
pytest>=8.3,<9.0
